requirement.txt
asgiref==3.7.2
attrs==24.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.44.1
crispy-bootstrap5==0.7
cryptography==41.0.6
defusedxml==0.7.1
dj-database-url==3.0.0
Django==4.2.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-heroku==0.3.1
django-taggit==4.0.0
django_csp==3.8
djangorestframework==3.14.0
gunicorn==23.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.3
MouseInfo==0.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==25.0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.7.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.2.0
pytz==2023.3
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.0.7
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
zipp==3.15.0
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.44.1
crispy-bootstrap5==0.7
cryptography==41.0.6
defusedxml==0.7.1
dj-database-url==3.0.0
Django==5.2.3
django-allauth==65.9.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==23.2
django-heroku==0.3.1
django-taggit==4.0.0
django_csp==3.8
djangorestframework==3.14.0
gunicorn==23.0.0
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
Markdown==3.4.3
MouseInfo==0.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==25.0
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyJWT==2.7.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-decouple==3.8
python3-openid==3.2.0
python3-xlib==0.15
pytweening==1.2.0
pytz==2023.3
PyYAML==6.0.2
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.27.1
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.4.4
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.0.7
webencodings==0.5.1
websocket-client==1.8.0
whitenoise==6.9.0
wsproto==1.2.0
zipp==3.15.0
